[wip] dockerfile builds image; runs container with docker-cmd command; https connection can be established; wss connection failes due to docker-assigned ips --- see description
Docker uses own Network namespaces:
When connecting via an ip-adress, a connection can be established to the running instance inside docker (especially when a port is specified). This works fine for the https-server, as the route can be resolved in the local network, when I type in the lan-ip-adress of my machine and connect to it via the https-protocol. The emerging problem is, that this app deeply relies on the usage of the websocket-protocol, for which we absolutely need to specify the (local)ip-context of the server, to which the client connects. This works fine and resolves nicely when running dockerless. However, when running a docker-container with the instance, the websocket-ip doesn't correctly resolve to either the incomming connection (172.17.0.2 or req.socket.remoteAddress), the server connection (172.17.0.1 or req.socket.localAddress), or localhost/0.0.0.0/127.0.0.1 .

These references seem to be promising:
* https://stackoverflow.com/questions/54101508/how-do-you-dockerize-a-websocket-server
* https://pythonspeed.com/articles/docker-connection-refused/